Exposed services
It works!!

+++ Spacelib/Spacelib-ejb/src/main/java/fr/miage/m1/facades/QuaiFacade.java
@@ -41,7 +41,7 @@ public class QuaiFacade extends AbstractFacade<Quai> implements QuaiFacadeLocal 
     }
 
     @Override
-    public Quai getQuai(int idQuai) {
+    public Quai getQuai(Long idQuai) {
         return this.find(idQuai);
     }
