+++ Spacelib/Spacelib-ejb/src/main/java/fr/miage/m1/facades/QuaiFacadeLocal.java
@@ -33,5 +33,5 @@ public interface QuaiFacadeLocal {
     
     public Quai creerQuai(int noQuai, boolean estLibre, Station station);
     
-    public Quai getQuai (int idQuai);
+    public Quai getQuai (Long idQuai);
 }
