+++ Spacelib/Spacelib-ejb/src/main/java/fr/miage/m1/metier/GestionQuaiLocal.java
@@ -18,5 +18,5 @@ public interface GestionQuaiLocal {
     
     public Quai creerQuai(int noQuai, boolean estLibre, Station station);
     
-    public Quai getQuai (int idQuai);
+    public Quai getQuai (Long idQuai);
 }
